+++ packages/patrol/lib/src/binding.dart
@@ -8,6 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:patrol/patrol.dart';
+import 'package:patrol/src/bs_coverage.dart';
 import 'package:patrol/src/devtools_service_extensions/devtools_service_extensions.dart';
 import 'package:patrol/src/global_state.dart' as global_state;
 import 'package:patrol/src/platform/current.dart' as current_platform;
@@ -101,6 +102,12 @@ class PatrolBinding extends LiveTestWidgetsFlutterBinding {
         final passed = global_state.isCurrentTestPassing;
         logger(
           'tearDown(): test "$testName" in group "$_currentDartTest", passed: $passed',
+        );
+
+        await BrowserStackCoverage.recordTestCompleted(
+          testName: testName,
+          testFilePath: _currentDartTest!,
+          passed: passed,
         );
 
         await patrolAppService.markDartTestAsCompleted(